Refactor : 여러 디렉토리로 컴포넌트 분리. props로 전달하지 않고 최종 사용하는 컴포넌트 내부에서 로직 실행하도록 정
