+++ src/Pages/AdminPage/AdminContainer/index.jsx
@@ -1,12 +1,10 @@
 import { useState, useEffect } from 'react';
-import { UserInfoService, VacationService, AllTableService } from 'API';
-import { API_PARAMS } from 'Utils/constants';
+import { UserInfoService } from 'API';
 
 import AdminModalButton from './AdminModalButton';
 import AdminTable from './AdminTable';
 import AdminChangeTable from './AdminChangeTable';
 import AdminModal from './AdminModal';
-import { addDays, format } from 'date-fns';
 
 const AdminContainer = ({ auth, userId, isOpen, setIsOpen }) => {
   const [users, setUsers] = useState([]);
@@ -15,8 +13,6 @@ const AdminContainer = ({ auth, userId, isOpen, setIsOpen }) => {
   const [tab, setTab] = useState(0);
   const [rowData, setRowData] = useState(null);
   const [selectRowData, setSelectRowData] = useState(null);
-  const [usersAttendence, setUsersAttendence] = useState([]);
-  const today = format(new Date(), 'yyyy-MM-dd');
 
   const updateSelectRowData = (curArrays, curTab) => {
     const filterArray = [];
@@ -28,31 +24,6 @@ const AdminContainer = ({ auth, userId, isOpen, setIsOpen }) => {
     });
 
     setSelectRowData(filterArray);
-  };
-
-  const handleChangeShuffleTeam = async (userId, team) => {
-    const result = await UserInfoService.patchTeam(userId, team);
-    getUser();
-  };
-
-  const handleAddVacation = async (select, addedDays) => {
-    const body = {
-      addedDays: addedDays,
-    };
-    const result = await VacationService.addVacation(select, body);
-    getUser();
-    setSelectUserId(null);
-  };
-  const handleMinusVacation = async (select, usedDays) => {
-    const body = { usedDays: usedDays, reason: 'Applied to all users.' };
-    const selectUser = rowData.filter(user => user.id === select);
-    if (selectUser[0].vacation === 0) {
-      // console.log('감소시킬 휴가가 없습니다!');
-      return;
-    }
-    const result = await VacationService.useVacation(select, body);
-    getUser();
-    setSelectUserId(null);
   };
 
   // const handleGetUser = async () => {
@@ -89,14 +60,8 @@ const AdminContainer = ({ auth, userId, isOpen, setIsOpen }) => {
     updateSelectRowData(newArray, tab);
   };
 
-  const getUserAttendance = async () => {
-    const result = await AllTableService.getTable(today);
-    setUsersAttendence(result.data);
-  };
-
   useEffect(() => {
     getUser();
-    getUserAttendance();
   }, []);
 
   return (
@@ -122,10 +87,8 @@ const AdminContainer = ({ auth, userId, isOpen, setIsOpen }) => {
         isOpen={isOpen}
         setIsOpen={setIsOpen}
         rowData={rowData}
-        usersAttendence={usersAttendence}
-        handleChangeShuffleTeam={handleChangeShuffleTeam}
-        handleAddVacation={handleAddVacation}
-        handleMinusVacation={handleMinusVacation}
+        getUser={getUser}
+        setSelectUserId={setSelectUserId}
       />
     </>
   );
